refactor: 퀴즈예외 패키지 이동
