+++ src/main/java/com/grepp/spring/app/model/quiz/service/QuizCreateService.java
@@ -13,7 +13,9 @@ import com.grepp.spring.app.model.quiz.repository.QuizRepository;
 import com.grepp.spring.app.model.quiz.repository.QuizSetRepository;
 import com.grepp.spring.app.model.study.entity.StudyGoal;
 import com.grepp.spring.app.model.study.repository.StudyGoalRepository;
-import com.grepp.spring.infra.error.exceptions.*;
+import com.grepp.spring.infra.error.exceptions.Quiz.QuizAlreadyExistsException;
+import com.grepp.spring.infra.error.exceptions.Quiz.QuizGenerationFailedException;
+import com.grepp.spring.infra.error.exceptions.Quiz.StudyGoalNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.retry.annotation.Backoff;
 import org.springframework.retry.annotation.Retryable;
