+++ src/main/java/com/grepp/spring/app/model/quiz/service/QuizGetService.java
@@ -5,8 +5,8 @@ import com.grepp.spring.app.model.quiz.dto.QuizDto;
 import com.grepp.spring.app.model.quiz.dto.QuizProjection;
 import com.grepp.spring.app.model.quiz.repository.QuizSetRepository;
 import com.grepp.spring.app.model.study.repository.StudyRepository;
-import com.grepp.spring.infra.error.exceptions.InvalidQuizException;
-import com.grepp.spring.infra.error.exceptions.MemberNotFoundException;
+import com.grepp.spring.infra.error.exceptions.Quiz.InvalidQuizException;
+import com.grepp.spring.infra.error.exceptions.Quiz.MemberNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
