+++ src/main/java/com/grepp/spring/app/model/quiz/service/SurvivalResultService.java
@@ -8,9 +8,9 @@ import com.grepp.spring.app.model.quiz.entity.QuizRecord;
 import com.grepp.spring.app.model.quiz.entity.QuizSet;
 import com.grepp.spring.app.model.quiz.repository.QuizRecordRepository;
 import com.grepp.spring.app.model.quiz.repository.QuizSetRepository;
-import com.grepp.spring.infra.error.exceptions.QuizResultAlreadySubmittedException;
-import com.grepp.spring.infra.error.exceptions.QuizSetNotFoundException;
-import com.grepp.spring.infra.error.exceptions.StudyMemberNotFoundException;
+import com.grepp.spring.infra.error.exceptions.Quiz.QuizResultAlreadySubmittedException;
+import com.grepp.spring.infra.error.exceptions.Quiz.QuizSetNotFoundException;
+import com.grepp.spring.infra.error.exceptions.Quiz.StudyMemberNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
