+++ src/main/java/com/grepp/spring/infra/error/exceptions/Quiz/InvalidQuizException.java
@@ -1,4 +1,4 @@
-package com.grepp.spring.infra.error.exceptions;
+package com.grepp.spring.infra.error.exceptions.Quiz;
 
 public class InvalidQuizException extends RuntimeException {
