+++ src/main/java/com/grepp/spring/infra/error/exceptions/Quiz/InvalidQuizGradeRequestException.java
@@ -1,4 +1,4 @@
-package com.grepp.spring.infra.error.exceptions;
+package com.grepp.spring.infra.error.exceptions.Quiz;
 
 public class InvalidQuizGradeRequestException extends RuntimeException  {
