+++ src/main/java/com/grepp/spring/infra/error/exceptions/Quiz/StudyGoalNotFoundException.java
@@ -1,4 +1,4 @@
-package com.grepp.spring.infra.error.exceptions;
+package com.grepp.spring.infra.error.exceptions.Quiz;
 
 public class StudyGoalNotFoundException extends RuntimeException  {
